refactor: create a slice to handle the state of posts
