+++ src/store.js
@@ -1,10 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
 
-import postReducer from './reducers/postReducer';
+import postsReducer from './slices/postsSlice';
 
 const store = configureStore({
   reducer: {
-    posts: postReducer,
+    posts: postsReducer,
   },
   devTools: true,
 });
